# docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact: {}
  description: B Y E C O M  is your ecom. shop solution, where you can sell anything
    online with ease and at an affordable price. We provide an eCommerce platform
    for businesses to sell their products online and connect with customers worldwide.
  title: B Y E C O M  LTD
  version: "1.0"
paths:
...
  /admin/product:
    delete:
      description: Deleting product completely
      parameters:
      - description: name search by id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Product Delete
      tags:
      - Admin Product
    get:
      description: Showing Products details in admin side
      produces:
      - application/json
      responses: {}
      summary: Products Show
      tags:
      - Admin Product
    post:
      consumes:
      - multipart/form-data
      description: Adding Product with it's details
      parameters:
      - description: Product Name
        in: formData
        name: name
        required: true
        type: string
      - description: Product Price
        in: formData
        name: price
        required: true
        type: integer
      - description: Product Offer
        in: formData
        name: offer
        required: true
        type: integer
      - collectionFormat: csv
        description: Product Color
        in: formData
        items:
          type: string
        name: color
        required: true
        type: array
      - description: Product Quantity
        in: formData
        name: quantity
        required: true
        type: integer
      - description: Product Description
        in: formData
        name: description
        required: true
        type: string
      - description: Product Category
        in: formData
        name: category
        required: true
        type: string
      - collectionFormat: csv
        description: Product Image
        in: formData
        items:
          type: file
        name: image
        required: true
        type: array
      produces:
      - application/json
      responses: {}
      summary: Product Add
      tags:
      - Admin Product
    put:
      consumes:
      - multipart/form-data
      description: Editing Product with it's details
      parameters:
      - description: name search by id
        in: query
        name: id
        required: true
        type: string
      - description: Product Name
        in: formData
        name: name
        required: true
        type: string
      - description: Product Price
        in: formData
        name: price
        required: true
        type: integer
      - collectionFormat: csv
        description: Product Color
        in: formData
        items:
          type: string
        name: color
        required: true
        type: array
      - description: Product Quantity
        in: formData
        name: quantity
        required: true
        type: integer
      - description: Product Description
        in: formData
        name: description
        required: true
        type: string
      - description: Product Category
        in: formData
        name: category
        required: true
        type: string
      - collectionFormat: csv
        description: Product Image
        in: formData
        items:
          type: file
        name: image
        required: true
        type: array
      produces:
      - application/json
      responses: {}
      summary: Product Edit
      tags:
      - Admin Product
...
  /user/cart:
    delete:
      description: Deleting a Product from Cart completely
      parameters:
      - description: name search by id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Cart Delete
      tags:
      - User Cart
    get:
      description: Showing Cart details in user side
      produces:
      - application/json
      responses: {}
      summary: Cart Show
      tags:
      - User Cart
    patch:
      description: Decreasing Cart quantity
      parameters:
      - description: cart id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Less Cart
      tags:
      - User Cart
    post:
      description: Adding Cart with it's details
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: string
      - description: product color
        in: query
        name: color
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Cart Add
      tags:
      - User Cart
...
  /user/filter:
    get:
      description: Filtering Products in user side
      parameters:
      - description: filter search
        in: query
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Filtered Products
      tags:
      - User Home&Product
...
  /user/search:
    get:
      description: Searching Products in user side
      parameters:
      - description: product search
        in: query
        name: search
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Searching Products
      tags:
      - User Home&Product
...
  /user/sort:
    get:
      description: Showing all Products details by sorting in user side
      parameters:
      - description: sort type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Sorted Home Show
      tags:
      - User Home&Product
...
swagger: "2.0"